Extract utils from Controller and View
